amogus

+++ include/odom.h
@@ -3,8 +3,8 @@
 
 const double pi = 3.141592653589793238462643;
 
-double angle360(double angle);
-double angle180(double angle);
+long double angle360(long double angle);
+long double angle180(long double angle);
 
 void starttracking(float o = 0);
-double getorientation();
+long double getorientation();
